+++ internal/embeddings/cache_test.go
@@ -1,5 +1,3 @@
-//go:build embeddings
-
 package embeddings
 
 import (
@@ -12,7 +10,7 @@ func TestEmbeddingCache_PutGet(t *testing.T) {
 	cache, err := newCache(100)
 	require.NoError(t, err)
 
-	embedding := []float32{0.1, 0.2, 0.3}
+	embedding := []float64{0.1, 0.2, 0.3}
 	cache.Put("test", embedding)
 
 	retrieved, ok := cache.Get("test")
@@ -33,9 +31,9 @@ func TestEmbeddingCache_Eviction(t *testing.T) {
 	cache, err := newCache(2)
 	require.NoError(t, err)
 
-	cache.Put("a", []float32{1})
-	cache.Put("b", []float32{2})
-	cache.Put("c", []float32{3}) // Should evict "a"
+	cache.Put("a", []float64{1})
+	cache.Put("b", []float64{2})
+	cache.Put("c", []float64{3}) // Should evict "a"
 
 	_, ok := cache.Get("a")
 	require.False(t, ok, "Expected 'a' to be evicted")
@@ -53,10 +51,10 @@ func TestEmbeddingCache_Size(t *testing.T) {
 
 	require.Equal(t, 0, cache.Size())
 
-	cache.Put("a", []float32{1})
+	cache.Put("a", []float64{1})
 	require.Equal(t, 1, cache.Size())
 
-	cache.Put("b", []float32{2})
+	cache.Put("b", []float64{2})
 	require.Equal(t, 2, cache.Size())
 }
 
@@ -64,8 +62,8 @@ func TestEmbeddingCache_Clear(t *testing.T) {
 	cache, err := newCache(100)
 	require.NoError(t, err)
 
-	cache.Put("a", []float32{1})
-	cache.Put("b", []float32{2})
+	cache.Put("a", []float64{1})
+	cache.Put("b", []float64{2})
 	require.Equal(t, 2, cache.Size())
 
 	cache.Clear()
@@ -76,7 +74,7 @@ func TestEmbeddingCache_CopyOnPut(t *testing.T) {
 	cache, err := newCache(100)
 	require.NoError(t, err)
 
-	original := []float32{1, 2, 3}
+	original := []float64{1, 2, 3}
 	cache.Put("test", original)
 
 	// Modify original
@@ -85,7 +83,7 @@ func TestEmbeddingCache_CopyOnPut(t *testing.T) {
 	// Retrieved should be unchanged
 	retrieved, ok := cache.Get("test")
 	require.True(t, ok)
-	require.Equal(t, float32(1), retrieved[0], "Cache should store a copy, not a reference")
+	require.Equal(t, float64(1), retrieved[0], "Cache should store a copy, not a reference")
 }
 
 func TestEmbeddingCache_DefaultSize(t *testing.T) {
@@ -99,8 +97,8 @@ func TestEmbeddingCache_Stats(t *testing.T) {
 	cache, err := newCache(100)
 	require.NoError(t, err)
 
-	cache.Put("a", []float32{1})
-	cache.Put("b", []float32{2})
+	cache.Put("a", []float64{1})
+	cache.Put("b", []float64{2})
 
 	stats := cache.Stats()
 	require.Equal(t, 2, stats.Size)
