+++ internal/embeddings/provider.go
@@ -1,5 +1,3 @@
-//go:build embeddings
-
 package embeddings
 
 import (
@@ -12,7 +10,7 @@ type EmbeddingProvider interface {
 	// Embed generates embeddings for a batch of texts.
 	// Returns L2-normalized vectors suitable for cosine similarity.
 	// The returned vectors MUST be L2-normalized.
-	Embed(ctx context.Context, texts []string) ([][]float32, error)
+	Embed(ctx context.Context, texts []string) ([][]float64, error)
 
 	// Dimension returns the embedding dimension for this provider.
 	Dimension() int
